move groupByTimeframe function to helper module
