+++ workers/loc.api/sync/get-risk/helpers/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const groupByTimeframe = require('./group-by-timeframe')
+
+module.exports = {
+  groupByTimeframe
+}
